+++ src/decoder-phrasebased/src/native/interpolated-lm/db/NGramBatch.h
@@ -29,6 +29,7 @@ namespace mmt {
 
         class NGramBatch {
             friend class NGramStorage;
+
         public:
 
             NGramBatch(uint8_t order, size_t maxSize) : NGramBatch(order, maxSize, vector<seqid_t>()) {}
@@ -66,7 +67,7 @@ namespace mmt {
             vector<memory_t> deletions;
 
             inline const bool ShouldAcceptUpdate(channel_t channel, seqid_t position) const {
-                return channel >= streams.size() || streams[channel] < position;
+                return channel >= (ssize_t) streams.size() || streams[channel] < position;
             }
         };
